+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - "Biology/Mammals/Walrus haulout area regulations" from NatureMap
   - "Biology/Mammals/Walrus wintering and spring areas" from NatureMap
   - "Biology/Mammals/Harbour seal areas" from NatureMap
+  - "Biology/Mammals/Beluga important areas" from NatureMap
 
 # v4.0.0alpha1 (2026-03-04)